requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
colorclass==2.2.2
decorator==5.1.1
docopt==0.6.2
frozenlist==1.3.0
fsspec==2022.2.0
gcsfs==2022.2.0
google-api-core==2.7.1
google-auth==1.32.1
google-auth-oauthlib==0.5.0
google-cloud-bigquery==2.34.2
google-cloud-bigquery-storage==2.13.0
google-cloud-core==2.2.3
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
joblib==1.1.0
multidict==6.0.2
numpy==1.22.0
oauthlib==3.2.0
packaging==21.3
pandas==1.3.5
pandas-gbq==0.14.1
pip-upgrader==1.4.15
proto-plus==1.20.3
protobuf==3.19.4
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
terminaltables==3.1.10
threadpoolctl==3.1.0
urllib3==1.26.8
yarl==1.7.2
